requirements-dev.txt
# Health API QA Framework - Development & Testing Dependencies
# Ferramentas para desenvolvimento, testes e qualidade de código

# Testing Framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-html==4.1.1
pytest-xdist==3.5.0
pytest-mock==3.12.0
pytest-benchmark==4.0.0
pytest-timeout==2.2.0

# Test Data & Fixtures
factory-boy==3.3.0
faker==20.1.0
freezegun==1.2.2

# API Testing
requests-mock==1.11.0
httpx-mock==0.7.0
responses==0.24.1

# Contract Testing
pact-python==2.2.1
dredd-hooks==0.2.0

# Performance Testing
locust==2.17.0
pytest-benchmark==4.0.0

# Security Testing
bandit==1.7.5
safety==2.3.5
semgrep==1.45.0

# Code Quality & Linting
black==23.11.0
isort==5.12.0
flake8==6.1.0
pylint==3.0.3
mypy==1.7.1
pre-commit==3.6.0

# Code Formatting
autopep8==2.0.4
yapf==0.40.2

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0
mkdocs==1.5.3
mkdocs-material==9.4.8

# Database Testing
pytest-postgresql==5.0.0
testcontainers==3.7.1

# Mocking & Stubbing
responses==0.24.1
vcrpy==5.1.0
betamax==0.8.1

# Load Testing
artillery==1.7.9

# API Documentation Testing
dredd==14.1.0
newman==5.3.2

# Reporting
allure-pytest==2.13.2
pytest-json-report==1.5.0
coverage==7.3.2

# Development Tools
ipython==8.17.2
jupyter==1.0.0
notebook==7.0.6

# Environment Management
python-dotenv==1.0.0
environs==10.0.0

# Debugging
pdb++==0.10.3
ipdb==0.13.13
pudb==2023.1

# Type Checking
types-requests==2.31.0.10
types-redis==4.6.0.11
types-PyYAML==6.0.12.12

# Async Testing
pytest-asyncio==0.21.1
aioresponses==0.7.4

# Database Migrations Testing
alembic==1.13.1

# Configuration Testing
dynaconf==3.2.4

# Time & Date Testing
freezegun==1.2.2
time-machine==2.13.0

# File System Testing
pyfakefs==5.3.2

# Network Testing
pytest-httpserver==1.0.8

# Parallel Testing
pytest-xdist==3.5.0
pytest-parallel==0.1.1

# Memory Profiling
memory-profiler==0.61.0
pympler==0.9

# Performance Monitoring
py-spy==0.3.14
